requirements.txt
Flask
requests
tinydb
Pillow